Because teams are not really user driven

+++ app/Http/Controllers/TeamsController.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace App\Http\Controllers\User;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
+use Exception;
+use App\Models\Team;
 use App\Http\Forms\TeamForm;
+use Illuminate\Http\Request;
+use App\Services\TeamService;
 use App\Http\Forms\TeamInviteForm;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\TeamCreateRequest;
 use App\Http\Requests\TeamUpdateRequest;
-use App\Models\Team;
-use App\Services\TeamService;
-use Exception;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Gate;
 
 class TeamsController extends Controller
 {
@@ -30,7 +31,7 @@ class TeamsController extends Controller
         $teams = $request->user()->teams;
         $memberships = $request->user()->teamMemberships;
 
-        return view('user.teams.index')->with(compact('memberships', 'teams'));
+        return view('teams.index')->with(compact('memberships', 'teams'));
     }
 
     /**
@@ -42,7 +43,7 @@ class TeamsController extends Controller
     {
         $form = app(TeamForm::class)->create();
 
-        return view('user.teams.create')->with(compact('form'));
+        return view('teams.create')->with(compact('form'));
     }
 
     /**
@@ -57,11 +58,11 @@ class TeamsController extends Controller
             $team = $this->service->create($request->except('_token'));
 
             if ($team) {
-                return redirect()->route('user.teams.edit', $team->id)
+                return redirect()->route('teams.edit', $team->id)
                     ->withMessage('Successfully created a team');
             }
 
-            return redirect()->route('user.teams')->withMessage('Failed to create a team');
+            return redirect()->route('teams')->withMessage('Failed to create a team');
         } catch (Exception $e) {
             return redirect()->back()->withErrors($e->getMessage());
         }
@@ -84,8 +85,27 @@ class TeamsController extends Controller
         $inviteForm = app(TeamInviteForm::class)
             ->setRoute('teams.members.invite', $team)->make();
 
-        return view('user.teams.edit')->with(compact('inviteForm', 'form', 'team'));
+        return view('teams.edit')->with(compact('inviteForm', 'form', 'team'));
     }
+
+    /**
+     * Show the form for handling members the specified team.
+     *
+     * @param  \App\Models\Team  $team
+     * @return \Illuminate\Http\Response
+     */
+    public function members(Team $team)
+    {
+        if (! Gate::allows('team-admin', $team)) {
+            abort(403);
+        }
+
+        $inviteForm = app(TeamInviteForm::class)
+            ->setRoute('teams.members.invite', $team)->make();
+
+        return view('teams.members')->with(compact('inviteForm', 'team'));
+    }
+
 
     /**
      * Update the specified team in storage.
@@ -96,8 +116,12 @@ class TeamsController extends Controller
      */
     public function update(TeamUpdateRequest $request, Team $team)
     {
+        if (Gate::denies('team-admin', $team)) {
+            return redirect()->back()->withErrors(['You do not have permission to do this.']);
+        }
+
         try {
-            if ($this->service->update($team, $request->except('_token'))) {
+            if ($this->service->update($team, $request)) {
                 return redirect()->back()->withMessage('Successfully updated');
             }
 
@@ -110,19 +134,52 @@ class TeamsController extends Controller
     /**
      * Remove the specified team from storage.
      *
-     * @param  int  $id
+     * @param  \App\Models\Team  $team
      * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Team $team)
     {
+        if (Gate::denies('team-admin', $team)) {
+            return redirect()->back()->withErrors(['You do not have permission to do this.']);
+        }
+
         try {
             $result = $this->service->destroy($team);
 
             if ($result) {
-                return redirect()->route('user.teams')->withMessage('Successfully deleted');
+                return redirect()->route('teams')->withMessage('Successfully deleted');
             }
 
-            return redirect()->route('user.teams')->withMessage('Failed to delete the team');
+            return redirect()->route('teams')->withMessage('Failed to delete the team');
+        } catch (Exception $e) {
+            return redirect()->back()->withErrors($e->getMessage());
+        }
+    }
+
+    /**
+     * Remove the specified team avatar from storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function destroyAvatar(Request $request)
+    {
+        $team = Team::find($request->team);
+
+        if (Gate::denies('team-admin', $team)) {
+            return redirect()->back()->withErrors(['You do not have permission to do this.']);
+        }
+
+        try {
+            Storage::delete($team->avatar);
+
+            if ($team->update([
+                'avatar' => null,
+            ])) {
+                return redirect()->back()->withMessage('Successfully deleted the team avatar.');
+            }
+
+            return redirect()->back()->withMessage('Failed to delete the team avatar.');
         } catch (Exception $e) {
             return redirect()->back()->withErrors($e->getMessage());
         }
